move dataplane mock tests into their own dir
