+++ packages/bsky/tests/data-plane/duplicate-records.test.ts
@@ -2,8 +2,8 @@ import { AtUri } from '@atproto/syntax'
 import { cidForCbor, TID } from '@atproto/common'
 import { WriteOpAction } from '@atproto/repo'
 import { TestNetwork } from '@atproto/dev-env'
-import * as lex from '../src/lexicon/lexicons'
-import { Database, PrimaryDatabase } from '../src'
+import * as lex from '../../src/lexicon/lexicons'
+import { Database, PrimaryDatabase } from '../../src'
 
 describe('duplicate record', () => {
   let network: TestNetwork
