+++ packages/bsky/tests/data-plane/indexing.test.ts
@@ -12,11 +12,11 @@ import AtpAgent, {
   AppBskyGraphFollow,
 } from '@atproto/api'
 import { TestNetwork, SeedClient } from '@atproto/dev-env'
-import { forSnapshot } from './_util'
-import usersSeed from './seeds/users'
-import basicSeed from './seeds/basic'
-import { ids } from '../src/lexicon/lexicons'
-import { Database } from '../src/data-plane/server/db'
+import { forSnapshot } from '../_util'
+import usersSeed from '../seeds/users'
+import basicSeed from '../seeds/basic'
+import { ids } from '../../src/lexicon/lexicons'
+import { Database } from '../../src/data-plane/server/db'
 
 describe('indexing', () => {
   let network: TestNetwork
